WebStorage/tests/submissions/Infraware/storage_session_length.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Web Storage</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>storage_session_length</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<script>
		if (('sessionStorage' in window) && window.sessionStorage !== null){
			sessionStorage.clear();
			
			test(function() {
				assert_equals(sessionStorage.length, 0, "sessionStorage.length")
			}, "sessionStorage length test after clear()")

			sessionStorage.setItem("name", "user1");
			sessionStorage.setItem("age", "20");

			test(function() {
				assert_equals(sessionStorage.length, 2, "sessionStorage.length")
			}, "sessionStorage length test after setItem()")

		} else {
			test(function()
			{
				assert_unreached("sessionStorage is not supported.");
			}, "sessionStorage support test")
		}	
	</script>

 </body>
</html>
